Support test filtering within Xcode
This fixes Google tests running when test execution was limited to a
different XCTest suite or case within Xcode. Rather than specifying a
test scope of only the suite/case to run Xcode specifies an inverse
filter excluding every test case except the selection. In order to
detect this the bundle defines a dummy testAll method and disables
running Google tests if this test case is excluded.