Support Xcode 7 by generating classes for each Google Test case
This was considered for prior implementations but avoided to simplify
the implementation and eliminate the risk of a generated class name
conflicting with an existing class. Originally XCTest's reporting and
Xcode integration supported this well, but in Xcode 6 parts of the
framework expected that test suite names corresponded to actual
Objective-C classes and test method names corresponded to instance
methods of those classes. Still, it was possible to work around this by
implementing undocumented XCTestCase methods that controlled the
reported class and method names for a test.

With Xcode 7 this workaround is no longer possible. The implementation
of test filtering has been made more efficient, but in a way that breaks
filtering compatibility with test suites added via +[XCTestCase
defaultTestSuite]. In prior releases the test set was built by locating
all XCTestCase subclasses, calling defaultTestSuite on them, then
filtering the results. Now the implementation assumes that it can pass
the test suite name to NSClassFromString() and directly obtain a class
already populated with methods for each test name in that suite. This is
not the case for test suites created via defaultTestSuite.

To continue to support filtering, treating Google Test cases as peers,
and to reduce conflicts with the XCTest framework, generate Objective-C
classes for each Google Test case and populate them with Objective-C
methods corresponding to each test in the Google Test case. The classes
are also given a custom implementation of +[XCTestCase testInvocations]
that returns test invocations for every method in the class.
Alternatively the generated methods could be prefixed with "test" to
allow XCTest to detect them automatically, but this could lead to
awkward test names if a Google Test name already starts with "test".

An alternative implementation for Xcode 7 compatibility would be to
sacrifice representing Google Tests as peers to standard XCTest suites
and instead represent them all as methods under an XCTestCase instance,
like "-[GoogleTests TestSuiteName.TestCaseName]". To support filtering
+[XCTestCase testCaseWithSelector:] could be overridden to recognize a
name following this pattern and return an appropriate XCTestCase
instance capable of running the test.

An additional note is that in Xcode 7 XCTest will instantiate an
instance of an NSPrincipalClass specified in the test bundle before
searching for tests. This could be used as an alternative to observing
the NSBundleDidLoadNotification for our own bundle.