Associate generated classes with the unit test bundle
This is accomplished by implementing the undocumented +bundleForClass
for the generated classes. This isn't necessary for proper operation,
but it results in more expected output from XCTest reports. With this
change Google Test classes appear to be part of the unit test bundle
that GoogleTests.mm is compiled into, as they did under earlier
implementations. Without it tests are still executed and filtered
properly, but appear in the report as part of a separate bundle for the
application running tests. When run within Xcode this is a command line
app so the bundle is represented as the name of the directory where that
app is located ("Agents").